requirement.txt
beautifulsoup4==4.6.0
requests==2.18.4
urllib3==1.22
selenium==3.8.0
mysql-connector-python==2.1.6
